add support for 64-bit ARM
Add a third processor type to the .so, and fix first compile-time
warnings and then a few dumb bugs based on assumptions about ptr sizes.
Works to play networked games and browse wordlists, but is not
extensively tested and needs to be before release.

+++ xwords4/android/app/src/main/java/org/eehouse/android/xw4/ThumbCanvas.java
@@ -54,7 +54,7 @@ public class ThumbCanvas extends BoardCanvas {
     // calling us.  So don't switch threads for the dict_getChars()
     // call
     @Override
-    public void dictChanged( int dictPtr )
+    public void dictChanged( long dictPtr )
     {
         if ( 0 != dictPtr ) {
             m_fontDims = null;
